third_party/WebKit/LayoutTests/fast/css/focus-ring-recursive-continuations.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('When adding focus ring rects, we should avoid adding duplicated rect for continuations');
onload = function() {
  document.body.offsetTop;
  // 6 focus ring rects:
  // - 0: 'SPAN0' part of the span;
  // - 1: div DIV1
  // - 2: first line box of div DIV1
  // - 3: div DIV2
  // - 4: first line box of div DIV2
  if (window.testRunner && window.internals)
    shouldBe("internals.focusRingRects(document.getElementById('focus')).length", "5");
};
</script>
<div>
  <span id="focus">SPAN0
    <div>DIV1
      <span>SPAN1
        <div>DIV2</div>
      </span>
    </div>
  </span>
</div>
